Move SASL auth interface outside impl
